requirements.txt
pyzmq==26.2.0
pydantic==2.9.1